Revert "vexpress64: Enable OF_CONTROL and OF_BOARD for VExpress64"
This patch only works for aarch64 (as the 'x' registers are not
available for ARMv7). Since this platform is ARMv7 in the previous
patch, this either needs to be changed or removed.  I opted to remove
it, as it doesn't seem to be necessary to boot the virtual hardware.
Given that the previous patch was rejected upstream, it is not
appropriate to fix this upstream.

Upstream-Status: Inappropriate
Signed-off-by: Jon Mason <jon.mason@arm.com>

This reverts commit 2661397464e47d45cd25bbc5e6b9de7594b3268d.